+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - **fix:** tests for validate.query
 - **rmv:** left console logs
 - **fix:** unset link header if no links are available
+- **add:** tests for header
 
 ### [v0.2.1] - 2016-06-26
 - **fix:** refactor file structure